Change the default install directory to remove the version string
Rework the deployable so that the non-osgi components will store the version information as part of the name. This avoids the scenario where redeploying the same application with a different version could undeploy the existing application but fail to deploy the new version. The new approach will enable the old version only when the new deployable successfully deploys.